Fix bug of null parentMap after submitting migration
In order to using the logics in the sub component of
form.dialog.component, passing the parentMap to the
samples.submitObject(), because the parentMap has been assigned even
when ngOnIint() is called. But it's null under "submitObject". The
statement parentMap[attribute.SYS_CODE] = {} is going to initialize
rather clear the content in the parentMap. In the form component, the
EntityFormInlineComponent will complain error " ERROR TypeError: Cannot
read property '594100e85707d6468b150a7a' of undefined", if the
parentMap[attribute.SYS_CODE] is not initialized appropriately, althogh
it'd be moved to the EntityFormInlineComponent logically.